+++ class12/ysj1224/Study/app/src/main/java/com/example/study/data/api/NaverApi.kt
@@ -17,6 +17,15 @@ interface NaverApi {
         @Header("X-Naver-Client-Secret") clientPw: String,
         @Query("query") query: String
     ): Call<NaverApiData>
+
+    object NaverRetrofit{
+        private val retrofit: Retrofit = Retrofit.Builder().baseUrl(baseUrl)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+        val SERVICE: NaverApi = retrofit.create(
+            NaverApi::class.java)
+    }
+
 }
