+++ class12/ysj1224/Study/app/src/main/java/com/example/study/data/remote/RemoteInterface.kt
@@ -1,0 +1,11 @@
+package com.example.study.data.remote
+
+import com.example.study.data.model.NaverApiData
+
+interface RemoteInterface {
+    fun getSearch(
+        query: String,
+        response: (List<NaverApiData.Item>) -> Unit,
+        fail: (Throwable) -> Unit
+    )
+}
