+++ class12/ysj1224/Study/app/src/main/java/com/example/study/data/repository/MovieListRepositoryImpl.kt
@@ -1,34 +1,18 @@
 package com.example.study.data.repository
 
 import com.example.study.data.model.NaverApiData
-import com.example.study.data.remote.RemoteService
-import com.example.study.view.clientId
-import com.example.study.view.clientSecret
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
+import com.example.study.data.remote.RemoteInterfaceImpl
 
 class MovieListRepositoryImpl : MovieListRepository {
+    private val remoteInterfaceImpl = RemoteInterfaceImpl()
+
     override fun doSearch(
         query: String,
         response: (List<NaverApiData.Item>) -> Unit,
         fail: (Throwable) -> Unit
     ) {
-        RemoteService.movieApiService.getSearch(
-            clientId = clientId,
-            clientPw = clientSecret,
-            query = query
-        ).enqueue(object : Callback<NaverApiData> {
-            override fun onFailure(call: Call<NaverApiData>, t: Throwable) {
-                fail(t)
-            }
-            override fun onResponse(call: Call<NaverApiData>, response: Response<NaverApiData>) {
-                if (response.isSuccessful) {
-                    response.body()?.let {
-                        response(it.items)
-                    }
-                }
-            }
-        })
+        remoteInterfaceImpl.getSearch(
+            query, response, fail
+        )
     }
 }
